aero-cli/pyproject.toml
[project]
name = "aero-cli"
version = "0.1.0"
description = "Command Center for the AeroRule ecosystem"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.12.0",
    "pydantic>=2.7.0",
    "openai>=1.12.0",
    "anthropic>=0.21.0",
    "google-generativeai>=0.4.0",
    "rich>=13.7.0",
    "cel-python",
    "pyyaml>=6.0.1",
    "termcolor>=2.4.0",
    "datamodel-code-generator>=0.25.0",
]

[project.scripts]
aero = "aerocli.main:app"

[tool.setuptools]
packages = ["aerocli"]

[tool.setuptools.package-data]
"aerocli" = ["*.json"]
